Map functionality improvements
The location of the feed table doesn’t jump around anymore, but stays
in selected feed location.
Instead of passing around a pointer to TAJLocationManager, a pointer to
a generic location provider protocol implementation is passed. Normally
it is TAJLocationManager, but when it’s map page, then it’s actually
TAJMapViewController (at the moment).
Added icon and launch screen.